backend/requirements.txt
flask
flask-socketio
eventlet
paho-mqtt
sqlalchemy
pymysql